+++ test/coordinator/e2e/coordinator/e2e_suite_test.go
@@ -136,7 +136,7 @@ var _ = ginkgo.BeforeSuite(func() {
 var _ = ginkgo.ReportAfterSuite("cleanup", func(report ginkgo.Report) {
 	if !report.SuiteSucceeded {
 		for idx := range numProcesses {
-			dumpPodsAndLogs(testutils.NamespaceForProcess(baseNsName, numProcesses, idx+1))
+			testutils.DumpPodsAndLogs(testConfig, testutils.NamespaceForProcess(baseNsName, numProcesses, idx+1))
 		}
 	}
